requirements.txt
pygame==2.6.1
torch==2.9.1
matplotlib==3.10.7
numpy==2.3.5
Pillow==12.0.0